requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-openai
google-api-python-client
python-dateutil
requests
